✨ files stucture managements
