+++ client/src/components/ProductReviews.tsx
@@ -11,7 +11,7 @@ import Typography from '@material-ui/core/Typography';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import Axios from 'axios';
-import { useStateValue } from './context/State.Context';
+import { useStateValue } from '../context/State.Context';
 import { Link } from 'react-router-dom';
 
 // Statics
