+++ client/src/pages/admin/Admin.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 import AdminProducts from './AdminProducts';
 import Divider from '@material-ui/core/Divider';
 import AdminOrders from './AdminOrders';
 import SettingsIcon from '@material-ui/icons/Settings';
 import Fab from '@material-ui/core/Fab';
 import AddIcon from '@material-ui/icons/Add';
-import { Link } from 'react-router-dom';
 
 // Statics
 import './Admin.css';
